--- lib/features/gallery_picker/presentation/pages/gallery_picker.dart
@@ -1,1 +1,0 @@
-
