+++ lib/features/gallery_picker/presentation/widgets/thumbnail_album.dart
