+++ lib/features/gallery_picker/presentation/widgets/thumbnail_media_file.dart
